+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.7.1] - 2018-05-13
+### Fixed
+
+- Changed how chapters are retrieved on Fanfox source. Fanfox is slowly removing the auxiliar javascript file with the chapter list and that change was breaking Foxy.
+
+## [0.7.0] - 2018-05-12
+### Added
+
+- New source: [MangaReader](https://www.mangareader.net).
+
+### Changed
+
+- Prevent cache during manga background updates.
+
+### Fixed
+
+- Kissmanga background updates. The updates were not working due to a change to the cloudflare code.
+
 ## [0.6.2] - 2018-03-12
 ### Added
 
@@ -154,7 +172,7 @@ This only affects users using the sync storage.
 - Settings page with Backup/Restore options.
 - License file.
 
-[Unreleased]: https://github.com/enakada/foxy-manga-reader/compare/v0.6.2...HEAD
+[Unreleased]: https://github.com/enakada/foxy-manga-reader/compare/v0.7.1...HEAD
 [0.2.0]: https://github.com/enakada/foxy-manga-reader/compare/v0.1.0...v0.2.0
 [0.3.0]: https://github.com/enakada/foxy-manga-reader/compare/v0.2.0...v0.3.0
 [0.3.1]: https://github.com/enakada/foxy-manga-reader/compare/v0.3.0...v0.3.1
@@ -164,4 +182,6 @@ This only affects users using the sync storage.
 [0.5.2]: https://github.com/enakada/foxy-manga-reader/compare/v0.5.1...v0.5.2
 [0.6.0]: https://github.com/enakada/foxy-manga-reader/compare/v0.5.2...v0.6.0
 [0.6.1]: https://github.com/enakada/foxy-manga-reader/compare/v0.6.0...v0.6.1
-[0.6.1]: https://github.com/enakada/foxy-manga-reader/compare/v0.6.1...v0.6.2
+[0.6.2]: https://github.com/enakada/foxy-manga-reader/compare/v0.6.1...v0.6.2
+[0.7.0]: https://github.com/enakada/foxy-manga-reader/compare/v0.6.2...v0.7.0
+[0.7.1]: https://github.com/enakada/foxy-manga-reader/compare/v0.7.0...v0.7.1